requirements.txt
asttokens==3.0.0
backcall==0.2.0
black==24.8.0
blobfile==2.1.1
certifi==2024.2.2
chardet==4.0.0
click==8.1.7
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
einops==0.8.0
executing==2.1.0
filelock==3.13.3
fonttools==4.53.0
fsspec==2024.10.0
h5py==3.11.0
huggingface-hub==0.26.2
idna==2.10
imageio==2.34.1
importlib_metadata==8.5.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.5
kmeans-pytorch==0.3
lxml==4.9.4
matplotlib==3.7.5
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.9.0.80
packaging==24.1
pandas==2.0.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.2
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodomex==3.20.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.2.0
regex==2024.11.6
requests==2.25.1
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.5.0
tokenizers==0.20.3
tomli==2.0.1
torch==1.13.1
torchattacks==3.5.1
torchvision==0.14.1
tornado==6.4.2
tqdm==4.66.2
traitlets==5.14.3
transformers==4.46.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
wcwidth==0.2.13
yacs==0.1.8
zipp==3.20.2
